pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "cloud-service-enum"
version = "2.0.0"
description = "Async-native multi-cloud enumeration tool aligned with CIS benchmarks (AWS, Azure, GCP) plus OSINT."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
authors = [{ name = "cloud-service-enum contributors" }]
keywords = ["aws", "azure", "gcp", "cloud", "security", "enumeration", "cis", "osint"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Security",
]

dependencies = [
    "click>=8.2",
    "rich>=14.0",
    "pydantic>=2.11",
    "pydantic-settings>=2.7",
    "httpx>=0.28",
    "anyio>=4.7",
]

[project.optional-dependencies]
aws = [
    "aioboto3>=15.0",
    "botocore>=1.40",
]
azure = [
    "azure-identity>=1.24",
    "azure-mgmt-resource>=24.0",
    "azure-mgmt-resource-policy>=1.0.0b2",
    "azure-mgmt-subscription>=3.1",
    "azure-mgmt-storage>=22.0",
    "azure-mgmt-compute>=34.0",
    "azure-mgmt-network>=28.0",
    "azure-mgmt-keyvault>=11.0",
    "azure-mgmt-sql>=3.0",
    "azure-mgmt-monitor>=6.0",
    "azure-mgmt-authorization>=4.0",
    "azure-mgmt-security>=7.0",
    "azure-mgmt-policyinsights>=1.0",
    "azure-mgmt-loganalytics>=13.0",
    "azure-mgmt-containerservice>=37.0",
    "azure-mgmt-containerregistry>=12.0",
    "azure-mgmt-cosmosdb>=9.0",
    "azure-mgmt-web>=8.0",
    "azure-mgmt-applicationinsights>=4.0",
    "azure-mgmt-eventhub>=11.0",
    "azure-mgmt-servicebus>=9.0",
    "azure-mgmt-eventgrid>=10.0",
    "azure-mgmt-apimanagement>=5.0",
    "azure-mgmt-frontdoor>=1.2",
    "azure-mgmt-cdn>=13.0",
    "azure-mgmt-logic>=10.0",
    "azure-mgmt-appcontainers>=4.0",
    "azure-mgmt-automation>=1.1.0b4",
    "azure-mgmt-rdbms>=10.2.0b18",
    "azure-mgmt-securityinsight>=2.0.0b2",
    "azure-mgmt-redhatopenshift>=2.0",
    "azure-storage-blob>=12.24",
    "msgraph-sdk>=1.56",
    "msrest>=0.7",
]
gcp = [
    "google-auth>=2.40",
    "google-api-python-client>=2.160",
    "google-cloud-resource-manager>=1.14",
    "google-cloud-iam>=2.18",
    "google-cloud-storage>=3.0",
    "google-cloud-compute>=1.27",
    "google-cloud-container>=2.50",
    "google-cloud-bigquery>=3.30",
    "google-cloud-kms>=3.4",
    "google-cloud-secret-manager>=2.23",
    "google-cloud-logging>=3.12",
    "google-cloud-monitoring>=2.27",
    "google-cloud-dns>=0.36",
    "google-cloud-functions>=1.20",
    "google-cloud-run>=0.10",
    "google-cloud-firestore>=2.20",
    "google-cloud-spanner>=3.51",
    "google-cloud-redis>=2.18",
    "google-cloud-securitycenter>=1.36",
    "google-cloud-aiplatform>=1.80",
    "google-cloud-pubsub>=2.27",
    "google-cloud-binary-authorization>=1.12",
]
reports = [
    "openpyxl>=3.1",
    "python-docx>=1.2",
    "fpdf2>=2.8",
    "lxml>=5.3",
]
osint = [
    "dnspython>=2.7",
    "python-whois>=0.9",
    "cryptography>=44.0",
]
all = [
    "cloud-service-enum[aws,azure,gcp,reports,osint]",
]

[project.scripts]
cse = "cloud_service_enum.cli:main"
cloud-service-enum = "cloud_service_enum.cli:main"

[project.urls]
Homepage = "https://github.com/example/cloud-service-enum"
Issues = "https://github.com/example/cloud-service-enum/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/cloud_service_enum"]

[tool.hatch.build.targets.wheel.force-include]
"src/cloud_service_enum/data" = "cloud_service_enum/data"
